# Stop OpenCL DeviceManager defaulting to CPU device