@nickgg 2020-01-21
@opti-mix fixed those, let me know what you think.